services/server/requirements.txt
aerich==0.6.3
aiosqlite==0.17.0
anyio==3.6.1
asyncpg==0.26.0
autopep8==1.6.0
click==8.1.3
dictdiffer==0.9.0
fastapi==0.78.0
h11==0.13.0
idna==3.3
iso8601==1.0.2
orjson==3.7.12
pep8==1.7.1
pycodestyle==2.8.0
pydantic==1.9.1
pypika-tortoise==0.1.5
pytz==2022.1
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
tomlkit==0.11.1
tortoise-orm==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
